LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/MediaQueryList-addListener-handleEvent-expected.txt


FAIL calls handleEvent method of event listener promise_test: Unhandled rejection with value: object "TypeError: Argument 1 ('listener') to MediaQueryList.addListener must be a function"
FAIL looks up handleEvent method on every event dispatch promise_test: Unhandled rejection with value: object "TypeError: Argument 1 ('listener') to MediaQueryList.addListener must be a function"
PASS doesn't look up handleEvent method on callable event listeners 
FAIL rethrows errors when getting handleEvent assert_equals: handleEvent property was not looked up expected 1 but got 0
FAIL throws if handleEvent is falsy and not callable assert_equals: handleEvent property was not looked up expected 1 but got 0
FAIL throws if handleEvent is thruthy and not callable assert_equals: handleEvent property was not looked up expected 1 but got 0
